Add new syscall & Fix memcpy, strlen & Add Makefile

+++ src/string/strlen.c
@@ -1,4 +1,6 @@
-int my_strlen(const unsigned char *str) {
+#include<ogc/stddef.h>
+
+size_t my_strlen(const unsigned char *str) {
   const unsigned char *s = str;
   
   while(*s) s++;
